Refactor codebase to support Vitest and remove integration tests
Various integration tests are removed from the codebase in a significant refactoring task. Along with this, the testing handler has been switched from 'CRA' to 'Vitest'. The aim of these changes is to improve dependency management and optimize our test strategies.
