+++ package.json
@@ -17,6 +17,7 @@
     "@types/react": "^18.2.60",
     "@types/react-dom": "^18.2.19",
     "@types/react-redux": "^7.1.33",
+    "@vitejs/plugin-react": "^4.2.1",
     "@vitejs/plugin-react-refresh": "^1.3.6",
     "axios": "^1.1.3",
     "react": "^18.2.0",
@@ -29,6 +30,7 @@
     "storybook-addon-react-router-v6": "^2.0.15",
     "styled-components": "^6.1.1",
     "typescript": "^5.3.3",
+    "vite-plugin-circular-dependency": "^0.4.1",
     "web-vitals": "^3.5.2"
   },
   "scripts": {
@@ -68,6 +70,7 @@
     "@storybook/react-webpack5": "^7.6.13",
     "@storybook/test": "^7.6.13",
     "@types/node": "^20.11.20",
+    "circular-dependency-plugin": "^5.2.2",
     "eslint": "8.22.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^8.7.0",
@@ -80,6 +83,7 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-storybook": "^0.6.15",
     "gh-pages": "^6.1.1",
+    "jsdom": "^24.0.0",
     "prettier": "^2.8.4",
     "prop-types": "^15.8.1",
     "puppeteer": "^22.0.0",
