+++ src/common/utils/index.ts
@@ -1,3 +1,3 @@
 export { handleServerAppError } from './handleServerAppError'
 export { handleServerNetworkError } from './handleServerNetworkError'
-export { createAppAsyncThunk } from './createAppAsyncThunk'
+//export { createAppAsyncThunk } from './createAppAsyncThunk'
