+++ src/features/tasks/Tasks.tsx
@@ -1,7 +1,7 @@
 import React, { memo } from 'react'
 import { NoItemsPrompt } from 'common/components'
 import List from '@mui/material/List'
-// import { useAutoAnimate } from '@formkit/auto-animate/react'
+import { useAutoAnimate } from '@formkit/auto-animate/react'
 import { useTasks } from './useTasks'
 import { Task } from './task/Task'
 
@@ -10,13 +10,13 @@ export type TasksProps = {
 }
 
 export const Tasks = memo(({ todolistId }: TasksProps) => {
-  // const [listRef] = useAutoAnimate<HTMLUListElement>()
+  const [listRef] = useAutoAnimate<HTMLUListElement>()
   const { tasks } = useTasks(todolistId)
 
   return (
     <>
       {tasks.length > 0 ? (
-        <List>
+        <List ref={listRef}>
           {tasks.map((task) => (
             <Task key={task.id} todolistId={todolistId} {...task} />
           ))}
